services/ui/requirements.txt
streamlit>=1.38
requests>=2.32
python-dotenv>=1.0
pydantic>=2.8
